Refactor to use Problem class
The previous verson of the code used a type parameter in optimizers,
which meant the specific solution subtype to be optimized must be known
at compile time. Now the subtype to optimize is a type field within
a problem class, so the optimizer can work an a problem of any kind
at runtime.